build(package-builder): adds jobs to the build workflow
These jobs are 'checkLicense' that check the validity of the Unity license use in this repository and the second one 'exportPackage' build the Unity package on all platform given (per default only windows 64 is given) and create a artifact store in the result of the action run.
This commit also add a 'rename.ps1' script that rename everything in the repository with the new project name given, this script will be use when a new project will be created with this template.